requirements.txt
Django==3.2.3
djangorestframework==3.12.4
psycopg2-binary==2.9.3
python-dotenv==1.0.1
